Module plat_egg::raw

Lower level egraph API

Modules

reflect_const

Types and traits for specify whether path compression is supported

semi_persistent1

One variant of semi_persistence

semi_persistent2

Another variant of semi_persistence

Structs

EGraphResidual

A RawEGraph without its classes that can be obtained by dereferencing a RawEGraph.

RawEClass

An equivalence class of enodes.

RawEGraph

A data structure to keep track of equalities between expressions.

RecExpr

A recursive expression from a user-defined Language.

UnionFind

Data structure that stores disjoint sets of Ids each with a representative

UnionInfo

Information about a call to RawEGraph::raw_union

Traits

AsUnwrap

Trait implemented for T and Option<T> used to provide bounds for push/pop impls

Language

Trait that defines a Language whose terms will be in the EGraph.

UndoLogPC

A sealed trait for types that can be used for push/pop APIs It is trivially implemented for ()

UndoLogT

A sealed trait for types that can be used for push/pop APIs It is trivially implemented for ()