Struct plat_egg::raw::EGraphResidual

pub struct EGraphResidual<L: Language> { /* private fields */ }

A RawEGraph without its classes that can be obtained by dereferencing a RawEGraph.

It exists as a separate type so that it can still be used while mutably borrowing a RawEClass

See RawEGraph::classes_mut, RawEGraph::get_class_mut

Implementations

impl<L: Language> EGraphResidual<L>

pub fn id_to_expr(&self, id: Id) -> RecExpr<L>

Pick a representative term for a given Id.

Calling this function on an uncanonical Id returns a representative based on how it was obtained

pub fn id_to_node(&self, id: Id) -> &L

Like id_to_expr but only goes one layer deep

pub fn find(&self, id: Id) -> Id

Canonicalizes an eclass id.

This corresponds to the find operation on the egraph’s underlying unionfind data structure.

Example

use plat_egg::{raw::*, SymbolLang as S};
let mut egraph = RawEGraph::<S, ()>::default();
let x = egraph.add_uncanonical(S::leaf("x"));
let y = egraph.add_uncanonical(S::leaf("y"));
assert_ne!(egraph.find(x), egraph.find(y));

egraph.union(x, y);
egraph.rebuild();
assert_eq!(egraph.find(x), egraph.find(y));

pub fn find_mut(&mut self, id: Id) -> Id

Same as find but requires mutable access since it does path compression

pub fn is_empty(&self) -> bool

Returns true if the egraph is empty

Example

use plat_egg::{raw::*, SymbolLang as S};
let mut egraph = RawEGraph::<S, ()>::default();
assert!(egraph.is_empty());
egraph.add_uncanonical(S::leaf("foo"));
assert!(!egraph.is_empty());

pub fn number_of_uncanonical_nodes(&self) -> usize

Returns the number of uncanonical enodes in the EGraph.

Example

use plat_egg::{raw::*, SymbolLang as S};
let mut egraph = RawEGraph::<S, ()>::default();
let x = egraph.add_uncanonical(S::leaf("x"));
let y = egraph.add_uncanonical(S::leaf("y"));
let fx = egraph.add_uncanonical(S::new("f", vec![x]));
let fy = egraph.add_uncanonical(S::new("f", vec![y]));
// only one eclass
egraph.union(x, y);
egraph.rebuild();

assert_eq!(egraph.number_of_uncanonical_nodes(), 4);
assert_eq!(egraph.number_of_classes(), 2);

pub fn uncanonical_nodes(
&self
) -> impl ExactSizeIterator<Item = (Id, &L)> + DoubleEndedIterator

Returns an iterator over the uncanonical ids in the egraph and the node that would be obtained by calling id_to_node on each of them

pub fn uncanonical_ids(&self) -> impl ExactSizeIterator<Item = Id> + 'static

Returns an iterator over all the uncanonical ids

pub fn total_size(&self) -> usize

Returns the number of enodes in the EGraph.

Actually returns the size of the hashcons index.

Example

use plat_egg::{raw::*, SymbolLang as S};
let mut egraph = RawEGraph::<S, ()>::default();
let x = egraph.add_uncanonical(S::leaf("x"));
let y = egraph.add_uncanonical(S::leaf("y"));
// only one eclass
egraph.union(x, y);
egraph.rebuild();

assert_eq!(egraph.total_size(), 2);
assert_eq!(egraph.number_of_classes(), 1);

pub fn lookup<B>(&self, enode: B) -> Option<Id> where
B: BorrowMut<L>,

Lookup the eclass of the given enode.

You can pass in either an owned enode or a &mut enode, in which case the enode’s children will be canonicalized.

Example

let mut egraph: RawEGraph<SymbolLang, ()> = Default::default();
let a = egraph.add_uncanonical(SymbolLang::leaf("a"));
let b = egraph.add_uncanonical(SymbolLang::leaf("b"));

// lookup will find this node if its in the egraph
let mut node_f_ab = SymbolLang::new("f", vec![a, b]);
assert_eq!(egraph.lookup(node_f_ab.clone()), None);
let id = egraph.add_uncanonical(node_f_ab.clone());
assert_eq!(egraph.lookup(node_f_ab.clone()), Some(id));

// if the query node isn't canonical, and its passed in by &mut instead of owned,
// its children will be canonicalized
egraph.union(a, b);
egraph.rebuild();
assert_eq!(egraph.lookup(&mut node_f_ab), Some(id));
assert_eq!(node_f_ab, SymbolLang::new("f", vec![a, a]));

pub fn lookup_uncanonical(&self, enode: &L) -> Option<Id>

Lookup the uncanonical id of the given uncanonical enode.

Returns Some(id) such that self.id_to_node(id) == enode if possible

Example

let mut egraph: RawEGraph<SymbolLang, ()> = Default::default();
let a = egraph.add_uncanonical(SymbolLang::leaf("a"));
let b = egraph.add_uncanonical(SymbolLang::leaf("b"));

let mut node_f_ab = SymbolLang::new("f", vec![a, b]);
let id = egraph.add_uncanonical(node_f_ab.clone());
assert_eq!(egraph.lookup_uncanonical(&node_f_ab), Some(id));

// if the query node isn't canonical, and its passed in by &mut instead of owned,
// its children will be canonicalized
egraph.union(a, b);
egraph.rebuild();
assert_eq!(egraph.lookup_uncanonical(&node_f_ab), Some(id));
assert_eq!(egraph.lookup_uncanonical(&SymbolLang::new("f", vec![a, a])), None);

pub fn lookup_expr(&self, expr: &RecExpr<L>) -> Option<Id>

Lookup the eclass of the given RecExpr.

Equivalent to the last value in EGraphResidual::lookup_expr_ids.

pub fn lookup_expr_ids(&self, expr: &RecExpr<L>) -> Option<Vec<Id>>

Lookup the eclasses of all the nodes in the given RecExpr.

pub fn generate_class_nodes(
&self
) -> IndexMap<Id, Vec<L>, BuildHasherDefault<FxHasher>>

Generate a mapping from canonical ids to the list of nodes they represent

pub fn dump_uncanonical(&self) -> impl Debug + '_

Returns a more debug-able representation of the egraph focusing on its uncanonical ids and nodes.

RawEGraphs implement Debug, but it’s not pretty. It prints a lot of stuff you probably don’t care about. This method returns a wrapper that implements Debug in a slightly nicer way, just dumping enodes in each eclass.

pub fn dot(&self) -> Dot<'_, L>

Creates a Dot to visualize this egraph. See Dot.