Trait plat_egg::raw::UndoLogPC

pub trait UndoLogPC {
    type AllowPathCompress: PathCompressT;
}

A sealed trait for types that can be used for push/pop APIs It is trivially implemented for ()

Required Associated Types

type AllowPathCompress: PathCompressT

When this type of undo log allows for path compression